feat: add trc10 and trc20 support
